+++ src/core/event-dispatchers/click-dispatcher.ts
@@ -1,8 +1,7 @@
 import { Injectable, Inject, NgZone, OnDestroy } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
-import { fromEvent, Subject, Observable, Observer, Subscription } from 'rxjs';
-import { auditTime } from 'rxjs/operators';
-import {ThyEventDispatcher} from './event-dispatcher';
+import { Observable } from 'rxjs';
+import { ThyEventDispatcher } from './event-dispatcher';
 
 const DEFAULT_CLICKED_TIME = 100;
 
@@ -10,11 +9,7 @@ const DEFAULT_CLICKED_TIME = 100;
     providedIn: 'root'
 })
 export class ThyClickDispatcher extends ThyEventDispatcher {
-
-    constructor(
-        @Inject(DOCUMENT) document: any,
-        ngZone: NgZone
-    ) {
+    constructor(@Inject(DOCUMENT) document: any, ngZone: NgZone) {
         super(document, ngZone, 'click');
     }