+++ src/dialog/test/dialog.spec.ts
@@ -12,7 +12,8 @@ import {
     fakeAsync,
     flushMicrotasks,
     inject,
-    flush
+    flush,
+    tick
 } from '@angular/core/testing';
 import { Location } from '@angular/common';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
@@ -27,8 +28,11 @@ import {
     DialogContentComponent,
     WithTemplateRefComponent,
     WithViewContainerDirective,
-    WithInjectedDataDialogComponent
+    WithInjectedDataDialogComponent,
+    WithOnPushViewContainerComponent
 } from './module';
+import { ESCAPE, A } from '../../util/keycodes';
+import { dispatchKeyboardEvent } from '../../core/testing';
 
 describe('ThyDialog', () => {
     let dialog: ThyDialog;
@@ -192,6 +196,17 @@ describe('ThyDialog', () => {
         expect(dialogContainerElement.getAttribute('role')).toBe('alertdialog');
     });
 
+    it('should pass initialState', () => {
+        const dialogRef = dialog.open(WithInjectedDataDialogComponent, {
+            initialState: {
+                data: `Hello initialState`
+            }
+        });
+        const instance = dialogRef.componentInstance;
+        viewContainerFixture.detectChanges();
+        expect(instance.data).toBe('Hello initialState');
+    });
+
     it('should close a dialog and get back a result', fakeAsync(() => {
         const dialogRef = dialog.open(DialogContentComponent, {
             viewContainerRef: testViewContainerRef
@@ -208,14 +223,136 @@ describe('ThyDialog', () => {
         expect(getDialogContainerElement()).toBeNull();
     }));
 
-    it('should pass initialState', () => {
-        const dialogRef = dialog.open(WithInjectedDataDialogComponent, {
-            initialState: {
-                data: `Hello initialState`
-            }
+    it('should close a dialog and get back a result before it is closed', fakeAsync(() => {
+        const dialogRef = dialog.open(DialogContentComponent, {
+            viewContainerRef: testViewContainerRef
         });
-        const instance = dialogRef.componentInstance;
+
+        flush();
         viewContainerFixture.detectChanges();
-        expect(instance.data).toBe('Hello initialState');
-    });
+
+        // beforeClose should emit before dialog container is destroyed
+        const beforeCloseHandler = jasmine
+            .createSpy('beforeClose callback')
+            .and.callFake(() => {
+                expect(getDialogContainerElement()).not.toBeNull(
+                    'dialog container exists when beforeClose is called'
+                );
+            });
+
+        dialogRef.beforeClosed().subscribe(beforeCloseHandler);
+        dialogRef.close('Bulbasaur');
+        viewContainerFixture.detectChanges();
+        flush();
+
+        expect(beforeCloseHandler).toHaveBeenCalledWith('Bulbasaur');
+        expect(getDialogContainerElement()).toBeNull();
+    }));
+
+    it('should close a dialog via the escape key', fakeAsync(() => {
+        dialog.open(DialogContentComponent, {
+            viewContainerRef: testViewContainerRef
+        });
+
+        dispatchKeyboardEvent(document.body, 'keydown', ESCAPE);
+        viewContainerFixture.detectChanges();
+        flush();
+
+        expect(getDialogContainerElement()).toBeNull();
+    }));
+
+    it('should close from a ViewContainerRef with OnPush change detection', fakeAsync(() => {
+        const onPushFixture = TestBed.createComponent(
+            WithOnPushViewContainerComponent
+        );
+
+        onPushFixture.detectChanges();
+
+        const dialogRef = dialog.open(DialogContentComponent, {
+            viewContainerRef: onPushFixture.componentInstance.viewContainerRef
+        });
+
+        flushMicrotasks();
+        onPushFixture.detectChanges();
+        flushMicrotasks();
+
+        expect(
+            overlayContainerElement.querySelectorAll('thy-dialog-container')
+                .length
+        ).toBe(1, 'Expected one open dialog.');
+
+        dialogRef.close();
+        flushMicrotasks();
+        onPushFixture.detectChanges();
+        tick(500);
+
+        expect(
+            overlayContainerElement.querySelectorAll('thy-dialog-container')
+                .length
+        ).toBe(0, 'Expected no open dialogs.');
+    }));
+
+    it('should close when clicking on the overlay backdrop', fakeAsync(() => {
+        dialog.open(DialogContentComponent, {
+            viewContainerRef: testViewContainerRef
+        });
+
+        viewContainerFixture.detectChanges();
+
+        const backdrop = overlayContainerElement.querySelector(
+            '.cdk-overlay-backdrop'
+        ) as HTMLElement;
+
+        backdrop.click();
+        viewContainerFixture.detectChanges();
+        flush();
+
+        expect(getDialogContainerElement()).toBeFalsy();
+    }));
+
+    it('should emit the backdropClick stream when clicking on the overlay backdrop', fakeAsync(() => {
+        const dialogRef = dialog.open(DialogContentComponent, {
+            viewContainerRef: testViewContainerRef
+        });
+
+        const spy = jasmine.createSpy('backdropClick spy');
+        dialogRef.backdropClick().subscribe(spy);
+
+        viewContainerFixture.detectChanges();
+
+        const backdrop = overlayContainerElement.querySelector(
+            '.cdk-overlay-backdrop'
+        ) as HTMLElement;
+
+        backdrop.click();
+        expect(spy).toHaveBeenCalledTimes(1);
+
+        viewContainerFixture.detectChanges();
+        flush();
+
+        // Additional clicks after the dialog has closed should not be emitted
+        backdrop.click();
+        expect(spy).toHaveBeenCalledTimes(1);
+    }));
+
+    it('should emit the keyboardEvent stream when key events target the overlay', fakeAsync(() => {
+        const dialogRef = dialog.open(DialogContentComponent, {
+            viewContainerRef: testViewContainerRef
+        });
+        tick(1);
+        const spy = jasmine.createSpy('keyboardEvent spy');
+        dialogRef.keydownEvents().subscribe(spy);
+
+        viewContainerFixture.detectChanges();
+
+        const backdrop = overlayContainerElement.querySelector(
+            '.cdk-overlay-backdrop'
+        ) as HTMLElement;
+        const container = getDialogContainerElement();
+        dispatchKeyboardEvent(document.body, 'keydown', A);
+        dispatchKeyboardEvent(document.body, 'keydown', A, backdrop);
+        dispatchKeyboardEvent(document.body, 'keydown', A, container);
+
+        expect(spy).toHaveBeenCalledTimes(3);
+    }));
 });