+++ src/grid/grid-column.component.ts
@@ -1,12 +1,25 @@
 import {
-    Component, Input, ElementRef, ViewEncapsulation, ContentChild,
-    ContentChildren, TemplateRef, ViewChild, InjectionToken, Optional, Inject
+    Component,
+    Input,
+    ElementRef,
+    ViewEncapsulation,
+    ContentChild,
+    ContentChildren,
+    TemplateRef,
+    ViewChild,
+    InjectionToken,
+    Optional,
+    Inject
 } from '@angular/core';
-import { AfterContentInit, OnInit, OnChanges, SimpleChanges } from '@angular/core';
+import {
+    AfterContentInit,
+    OnInit,
+    OnChanges,
+    SimpleChanges
+} from '@angular/core';
 import { inputValueToBoolean, isUndefined } from '../util/helpers';
 import { ThyGridComponent } from './grid.component';
 import { ThyGridColumn } from './grid.interface';
-
 
 export interface IThyGridColumnParentComponent {
     updateColumnSelections(key: string, selections: any): void;
@@ -15,8 +28,9 @@ export interface IThyGridColumnParentComponent {
 /**
  * Injection token used to provide the parent component to options.
  */
-export const THY_GRID_COLUMN_PARENT_COMPONENT =
-    new InjectionToken<IThyGridColumnParentComponent>('THY_GRID_COLUMN_PARENT_COMPONENT');
+export const THY_GRID_COLUMN_PARENT_COMPONENT = new InjectionToken<
+    IThyGridColumnParentComponent
+>('THY_GRID_COLUMN_PARENT_COMPONENT');
 
 @Component({
     selector: 'thy-grid-column',
@@ -24,7 +38,6 @@ export const THY_GRID_COLUMN_PARENT_COMPONENT =
     encapsulation: ViewEncapsulation.None
 })
 export class ThyGridColumnComponent implements OnInit {
-
     @Input() thyModelKey = '';
 
     @Input() thyTitle = '';
@@ -55,9 +68,18 @@ export class ThyGridColumnComponent implements OnInit {
 
     @Input() thyDefaultText = '';
 
-    @ContentChild(TemplateRef) templateRef: TemplateRef<any>;
+    @ContentChild('header') headerTemplateRef: TemplateRef<any>;
 
     @ContentChild('cell') cellTemplateRef: TemplateRef<any>;
+
+    @ContentChild(TemplateRef)
+    set templateRef(value: TemplateRef<any>) {
+        if (value) {
+            if (!this.headerTemplateRef && !this.cellTemplateRef) {
+                this.cellTemplateRef = value;
+            }
+        }
+    }
 
     public key?: string;
 
@@ -83,9 +105,10 @@ export class ThyGridColumnComponent implements OnInit {
 
     constructor(
         private el: ElementRef,
-        @Optional() @Inject(THY_GRID_COLUMN_PARENT_COMPONENT) public parent: IThyGridColumnParentComponent
-    ) {
-    }
+        @Optional()
+        @Inject(THY_GRID_COLUMN_PARENT_COMPONENT)
+        public parent: IThyGridColumnParentComponent
+    ) {}
 
     ngOnInit() {
         this.key = this._generateKey();
@@ -97,11 +120,15 @@ export class ThyGridColumnComponent implements OnInit {
         this.headerClassName = this.thyHeaderClassName;
         this.disabled = this.thyDisabled;
         this.defaultText = this.thyDefaultText;
-        this.templateRef = this.templateRef || this.cellTemplateRef;
         this._firstChange = false;
     }
 
     private _generateKey() {
-        return '[$$column]' + Math.random().toString(16).substr(2, 8);
+        return (
+            '[$$column]' +
+            Math.random()
+                .toString(16)
+                .substr(2, 8)
+        );
     }
 }